initial frontend changes and shared google refactoring
