+++ backend/danswer/server/documents/connector.py
@@ -23,35 +23,38 @@ from danswer.background.celery.versioned_apps.primary import app as primary_app
 from danswer.configs.app_configs import ENABLED_CONNECTOR_TYPES
 from danswer.configs.constants import DocumentSource
 from danswer.configs.constants import FileOrigin
-from danswer.connectors.gmail.connector_auth import delete_gmail_service_account_key
-from danswer.connectors.gmail.connector_auth import delete_google_app_gmail_cred
-from danswer.connectors.gmail.connector_auth import get_gmail_auth_url
-from danswer.connectors.gmail.connector_auth import get_gmail_service_account_key
-from danswer.connectors.gmail.connector_auth import get_google_app_gmail_cred
-from danswer.connectors.gmail.connector_auth import (
-    update_gmail_credential_access_tokens,
+from danswer.connectors.cross_connector_utils.google.google_auth import (
+    get_google_oauth_creds,
 )
-from danswer.connectors.gmail.connector_auth import (
-    upsert_gmail_service_account_key,
+from danswer.connectors.cross_connector_utils.google.google_kv import (
+    build_service_account_creds,
 )
-from danswer.connectors.gmail.connector_auth import upsert_google_app_gmail_cred
-from danswer.connectors.google_drive.connector_auth import build_service_account_creds
-from danswer.connectors.google_drive.connector_auth import DB_CREDENTIALS_DICT_TOKEN_KEY
-from danswer.connectors.google_drive.connector_auth import delete_google_app_cred
-from danswer.connectors.google_drive.connector_auth import delete_service_account_key
-from danswer.connectors.google_drive.connector_auth import get_auth_url
-from danswer.connectors.google_drive.connector_auth import get_google_app_cred
-from danswer.connectors.google_drive.connector_auth import (
-    get_google_drive_creds_for_authorized_user,
+from danswer.connectors.cross_connector_utils.google.google_kv import (
+    delete_google_app_cred,
 )
-from danswer.connectors.google_drive.connector_auth import get_service_account_key
-from danswer.connectors.google_drive.connector_auth import GOOGLE_DRIVE_SCOPES
-from danswer.connectors.google_drive.connector_auth import (
+from danswer.connectors.cross_connector_utils.google.google_kv import (
+    delete_service_account_key,
+)
+from danswer.connectors.cross_connector_utils.google.google_kv import get_auth_url
+from danswer.connectors.cross_connector_utils.google.google_kv import (
+    get_google_app_cred,
+)
+from danswer.connectors.cross_connector_utils.google.google_kv import (
+    get_service_account_key,
+)
+from danswer.connectors.cross_connector_utils.google.google_kv import (
     update_credential_access_tokens,
 )
-from danswer.connectors.google_drive.connector_auth import upsert_google_app_cred
-from danswer.connectors.google_drive.connector_auth import upsert_service_account_key
-from danswer.connectors.google_drive.connector_auth import verify_csrf
+from danswer.connectors.cross_connector_utils.google.google_kv import (
+    upsert_google_app_cred,
+)
+from danswer.connectors.cross_connector_utils.google.google_kv import (
+    upsert_service_account_key,
+)
+from danswer.connectors.cross_connector_utils.google.google_kv import verify_csrf
+from danswer.connectors.cross_connector_utils.google.shared_constants import (
+    DB_CREDENTIALS_DICT_TOKEN_KEY,
+)
 from danswer.db.connector import create_connector
 from danswer.db.connector import delete_connector
 from danswer.db.connector import fetch_connector_by_id
@@ -125,7 +128,7 @@ def check_google_app_gmail_credentials_exist(
     _: User = Depends(current_curator_or_admin_user),
 ) -> dict[str, str]:
     try:
-        return {"client_id": get_google_app_gmail_cred().web.client_id}
+        return {"client_id": get_google_app_cred(DocumentSource.GMAIL).web.client_id}
     except KvKeyNotFoundError:
         raise HTTPException(status_code=404, detail="Google App Credentials not found")
 
@@ -135,7 +138,7 @@ def upsert_google_app_gmail_credentials(
     app_credentials: GoogleAppCredentials, _: User = Depends(current_admin_user)
 ) -> StatusResponse:
     try:
-        upsert_google_app_gmail_cred(app_credentials)
+        upsert_google_app_cred(app_credentials, DocumentSource.GMAIL)
     except ValueError as e:
         raise HTTPException(status_code=400, detail=str(e))
 
@@ -150,7 +153,7 @@ def delete_google_app_gmail_credentials(
     db_session: Session = Depends(get_session),
 ) -> StatusResponse:
     try:
-        delete_google_app_gmail_cred()
+        delete_google_app_cred(DocumentSource.GMAIL)
         cleanup_gmail_credentials(db_session=db_session)
     except KvKeyNotFoundError as e:
         raise HTTPException(status_code=400, detail=str(e))
@@ -165,7 +168,9 @@ def check_google_app_credentials_exist(
     _: User = Depends(current_curator_or_admin_user),
 ) -> dict[str, str]:
     try:
-        return {"client_id": get_google_app_cred().web.client_id}
+        return {
+            "client_id": get_google_app_cred(DocumentSource.GOOGLE_DRIVE).web.client_id
+        }
     except KvKeyNotFoundError:
         raise HTTPException(status_code=404, detail="Google App Credentials not found")
 
@@ -175,7 +180,7 @@ def upsert_google_app_credentials(
     app_credentials: GoogleAppCredentials, _: User = Depends(current_admin_user)
 ) -> StatusResponse:
     try:
-        upsert_google_app_cred(app_credentials)
+        upsert_google_app_cred(app_credentials, DocumentSource.GOOGLE_DRIVE)
     except ValueError as e:
         raise HTTPException(status_code=400, detail=str(e))
 
@@ -190,7 +195,7 @@ def delete_google_app_credentials(
     db_session: Session = Depends(get_session),
 ) -> StatusResponse:
     try:
-        delete_google_app_cred()
+        delete_google_app_cred(DocumentSource.GOOGLE_DRIVE)
         cleanup_google_drive_credentials(db_session=db_session)
     except KvKeyNotFoundError as e:
         raise HTTPException(status_code=400, detail=str(e))
@@ -205,7 +210,11 @@ def check_google_service_gmail_account_key_exist(
     _: User = Depends(current_curator_or_admin_user),
 ) -> dict[str, str]:
     try:
-        return {"service_account_email": get_gmail_service_account_key().client_email}
+        return {
+            "service_account_email": get_service_account_key(
+                DocumentSource.GMAIL
+            ).client_email
+        }
     except KvKeyNotFoundError:
         raise HTTPException(
             status_code=404, detail="Google Service Account Key not found"
@@ -217,7 +226,7 @@ def upsert_google_service_gmail_account_key(
     service_account_key: GoogleServiceAccountKey, _: User = Depends(current_admin_user)
 ) -> StatusResponse:
     try:
-        upsert_gmail_service_account_key(service_account_key)
+        upsert_service_account_key(service_account_key, DocumentSource.GMAIL)
     except ValueError as e:
         raise HTTPException(status_code=400, detail=str(e))
 
@@ -232,7 +241,7 @@ def delete_google_service_gmail_account_key(
     db_session: Session = Depends(get_session),
 ) -> StatusResponse:
     try:
-        delete_gmail_service_account_key()
+        delete_service_account_key(DocumentSource.GMAIL)
         cleanup_gmail_credentials(db_session=db_session)
     except KvKeyNotFoundError as e:
         raise HTTPException(status_code=400, detail=str(e))
@@ -247,7 +256,11 @@ def check_google_service_account_key_exist(
     _: User = Depends(current_curator_or_admin_user),
 ) -> dict[str, str]:
     try:
-        return {"service_account_email": get_service_account_key().client_email}
+        return {
+            "service_account_email": get_service_account_key(
+                DocumentSource.GOOGLE_DRIVE
+            ).client_email
+        }
     except KvKeyNotFoundError:
         raise HTTPException(
             status_code=404, detail="Google Service Account Key not found"
@@ -259,7 +272,7 @@ def upsert_google_service_account_key(
     service_account_key: GoogleServiceAccountKey, _: User = Depends(current_admin_user)
 ) -> StatusResponse:
     try:
-        upsert_service_account_key(service_account_key)
+        upsert_service_account_key(service_account_key, DocumentSource.GOOGLE_DRIVE)
     except ValueError as e:
         raise HTTPException(status_code=400, detail=str(e))
 
@@ -274,7 +287,7 @@ def delete_google_service_account_key(
     db_session: Session = Depends(get_session),
 ) -> StatusResponse:
     try:
-        delete_service_account_key()
+        delete_service_account_key(DocumentSource.GOOGLE_DRIVE)
         cleanup_google_drive_credentials(db_session=db_session)
     except KvKeyNotFoundError as e:
         raise HTTPException(status_code=400, detail=str(e))
@@ -322,7 +335,7 @@ def upsert_gmail_service_account_credential(
     try:
         credential_base = build_service_account_creds(
             DocumentSource.GMAIL,
-            primary_admin_email=service_account_credential_request.gmail_delegated_user,
+            primary_admin_email=service_account_credential_request.gmail_primary_admin,
         )
     except KvKeyNotFoundError as e:
         raise HTTPException(status_code=400, detail=str(e))
@@ -349,9 +362,9 @@ def check_drive_tokens(
     ):
         return AuthStatus(authenticated=False)
     token_json_str = str(db_credentials.credential_json[DB_CREDENTIALS_DICT_TOKEN_KEY])
-    google_drive_creds = get_google_drive_creds_for_authorized_user(
+    google_drive_creds = get_google_oauth_creds(
         token_json_str=token_json_str,
-        scopes=GOOGLE_DRIVE_SCOPES,
+        source=DocumentSource.GOOGLE_DRIVE,
     )
     if google_drive_creds is None:
         return AuthStatus(authenticated=False)
@@ -882,7 +895,7 @@ def gmail_auth(
         httponly=True,
         max_age=600,
     )
-    return AuthUrl(auth_url=get_gmail_auth_url(int(credential_id)))
+    return AuthUrl(auth_url=get_auth_url(int(credential_id)))
 
 
 @router.get("/connector/google-drive/authorize/{credential_id}")
@@ -914,8 +927,8 @@ def gmail_callback(
     credential_id = int(credential_id_cookie)
     verify_csrf(credential_id, callback.state)
     if (
-        update_gmail_credential_access_tokens(
-            callback.code, credential_id, user, db_session
+        update_credential_access_tokens(
+            callback.code, credential_id, user, db_session, DocumentSource.GMAIL
         )
         is None
     ):
@@ -942,7 +955,7 @@ def google_drive_callback(
     verify_csrf(credential_id, callback.state)
 
     credentials: Credentials | None = update_credential_access_tokens(
-        callback.code, credential_id, user, db_session
+        callback.code, credential_id, user, db_session, DocumentSource.GOOGLE_DRIVE
     )
     if credentials is None:
         raise HTTPException(
