+++ backend/tests/daily/connectors/google_drive/test_google_drive_slim_docs.py
@@ -4,8 +4,10 @@ from unittest.mock import MagicMock
 from unittest.mock import patch
 
 from danswer.access.models import ExternalAccess
+from danswer.connectors.cross_connector_utils.google.google_utils import (
+    execute_paginated_retrieval,
+)
 from danswer.connectors.google_drive.connector import GoogleDriveConnector
-from danswer.connectors.google_drive.google_utils import execute_paginated_retrieval
 from ee.danswer.external_permissions.google_drive.doc_sync import (
     _get_permissions_from_slim_doc,
 )
