+++ web/src/app/admin/connectors/[connector]/pages/gmail/Credential.tsx
@@ -10,7 +10,8 @@ import { GMAIL_AUTH_IS_ADMIN_COOKIE_NAME } from "@/lib/constants";
 import Cookies from "js-cookie";
 import { TextFormField } from "@/components/admin/connectors/Field";
 import { Form, Formik } from "formik";
-import { Card } from "@tremor/react";
+import { User } from "@/lib/types";
+
 import {
   Credential,
   GmailCredentialJson,
@@ -293,9 +294,10 @@ interface DriveCredentialSectionProps {
   setPopup: (popupSpec: PopupSpec | null) => void;
   refreshCredentials: () => void;
   connectorExists: boolean;
+  user: User | null;
 }
 
-export const GmailOAuthSection = ({
+export const GmailAuthSection = ({
   gmailPublicCredential,
   gmailServiceAccountCredential,
   serviceAccountKeyData,
@@ -303,6 +305,7 @@ export const GmailOAuthSection = ({
   setPopup,
   refreshCredentials,
   connectorExists,
+  user,
 }: DriveCredentialSectionProps) => {
   const router = useRouter();
 
@@ -342,80 +345,76 @@ export const GmailOAuthSection = ({
     return (
       <div>
         <p className="text-sm mb-2">
-          When using a Gmail Service Account, you can either have Danswer act as
-          the service account itself OR you can specify an account for the
-          service account to impersonate.
+          When using a Gmail Service Account, you must specify the email of the
+          primary admin that you would like the service account to impersonate.
           <br />
           <br />
-          If you want to use the service account itself, leave the{" "}
-          <b>&apos;User email to impersonate&apos;</b> field blank when
-          submitting. If you do choose this option, make sure you have shared
-          the documents you want to index with the service account.
+          For this connector to index all users Gmail, the primary admin email
+          should be an owner/admin of the Google Organization that being
+          indexed.
         </p>
 
-        <Card>
-          <Formik
-            initialValues={{
-              gmail_delegated_user: "",
-            }}
-            validationSchema={Yup.object().shape({
-              gmail_delegated_user: Yup.string().optional(),
-            })}
-            onSubmit={async (values, formikHelpers) => {
-              formikHelpers.setSubmitting(true);
+        <Formik
+          initialValues={{
+            gmail_primary_admin: user?.email || "",
+          }}
+          validationSchema={Yup.object().shape({
+            gmail_primary_admin: Yup.string().required(),
+          })}
+          onSubmit={async (values, formikHelpers) => {
+            formikHelpers.setSubmitting(true);
 
-              const response = await fetch(
-                "/api/manage/admin/connector/gmail/service-account-credential",
-                {
-                  method: "PUT",
-                  headers: {
-                    "Content-Type": "application/json",
-                  },
-                  body: JSON.stringify({
-                    gmail_delegated_user: values.gmail_delegated_user,
-                  }),
-                }
-              );
-
-              if (response.ok) {
-                setPopup({
-                  message: "Successfully created service account credential",
-                  type: "success",
-                });
-              } else {
-                const errorMsg = await response.text();
-                setPopup({
-                  message: `Failed to create service account credential - ${errorMsg}`,
-                  type: "error",
-                });
+            const response = await fetch(
+              "/api/manage/admin/connector/gmail/service-account-credential",
+              {
+                method: "PUT",
+                headers: {
+                  "Content-Type": "application/json",
+                },
+                body: JSON.stringify({
+                  gmail_primary_admin: values.gmail_primary_admin,
+                }),
               }
-              refreshCredentials();
-            }}
-          >
-            {({ isSubmitting }) => (
-              <Form>
-                <TextFormField
-                  name="gmail_delegated_user"
-                  label="[Optional] User email to impersonate:"
-                  subtext="If left blank, Danswer will use the service account itself."
-                />
-                <div className="flex">
-                  <button
-                    type="submit"
-                    disabled={isSubmitting}
-                    className={
-                      "bg-slate-500 hover:bg-slate-700 text-white " +
-                      "font-bold py-2 px-4 rounded focus:outline-none " +
-                      "focus:shadow-outline w-full max-w-sm mx-auto"
-                    }
-                  >
-                    Submit
-                  </button>
-                </div>
-              </Form>
-            )}
-          </Formik>
-        </Card>
+            );
+
+            if (response.ok) {
+              setPopup({
+                message: "Successfully created service account credential",
+                type: "success",
+              });
+            } else {
+              const errorMsg = await response.text();
+              setPopup({
+                message: `Failed to create service account credential - ${errorMsg}`,
+                type: "error",
+              });
+            }
+            refreshCredentials();
+          }}
+        >
+          {({ isSubmitting }) => (
+            <Form>
+              <TextFormField
+                name="gmail_primary_admin"
+                label="Primary Admin Email:"
+                subtext="If left blank, Danswer will use the service account itself."
+              />
+              <div className="flex">
+                <button
+                  type="submit"
+                  disabled={isSubmitting}
+                  className={
+                    "bg-slate-500 hover:bg-slate-700 text-white " +
+                    "font-bold py-2 px-4 rounded focus:outline-none " +
+                    "focus:shadow-outline w-full max-w-sm mx-auto"
+                  }
+                >
+                  Submit
+                </button>
+              </div>
+            </Form>
+          )}
+        </Formik>
       </div>
     );
   }
