+++ web/src/app/admin/connectors/[connector]/pages/gmail/GmailPage.tsx
@@ -10,7 +10,7 @@ import {
   GmailCredentialJson,
   GmailServiceAccountCredentialJson,
 } from "@/lib/connectors/credentials";
-import { GmailOAuthSection, GmailJsonUploadSection } from "./Credential";
+import { GmailAuthSection, GmailJsonUploadSection } from "./Credential";
 import {
   usePublicCredentials,
   useConnectorCredentialIndexingStatus,
@@ -20,7 +20,7 @@ import { GmailConfig } from "@/lib/connectors/connectors";
 import { useUser } from "@/components/user/UserProvider";
 
 export const GmailMain = () => {
-  const { isLoadingUser, isAdmin } = useUser();
+  const { isLoadingUser, isAdmin, user } = useUser();
 
   const {
     data: appCredentialData,
@@ -142,7 +142,7 @@ export const GmailMain = () => {
           <Title className="mb-2 mt-6 ml-auto mr-auto">
             Step 2: Authenticate with Danswer
           </Title>
-          <GmailOAuthSection
+          <GmailAuthSection
             setPopup={setPopup}
             refreshCredentials={refreshCredentials}
             gmailPublicCredential={gmailPublicCredential}
@@ -150,6 +150,7 @@ export const GmailMain = () => {
             appCredentialData={appCredentialData}
             serviceAccountKeyData={serviceAccountKeyData}
             connectorExists={gmailConnectorIndexingStatuses.length > 0}
+            user={user}
           />
         </>
       )}
